# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupportDesk API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  workspaceName: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface InviteInput {
  name: string;
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  workspaceId: number;
  isOnline: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Workspace {
  id: number;
  name: string;
  ownerId: number;
  widgetColor: string;
  greetingMessage: string;
  botName: string;
  plan: string;
  /** @nullable */
  stripeCustomerId?: string | null;
  /** @nullable */
  stripeSubscriptionId?: string | null;
  createdAt: string;
}

export interface WorkspaceUpdate {
  name?: string;
  widgetColor?: string;
  greetingMessage?: string;
  botName?: string;
}

export interface PublicWorkspace {
  id: number;
  name: string;
  widgetColor: string;
  greetingMessage: string;
  botName: string;
  agentsOnline: boolean;
}

export interface Conversation {
  id: number;
  workspaceId: number;
  visitorName: string;
  visitorEmail: string;
  sourcePage: string;
  /** @nullable */
  browser?: string | null;
  status: string;
  /** @nullable */
  assignedAgentId?: number | null;
  assignedAgent?: User | null;
  createdAt: string;
  updatedAt?: string;
  unreadCount: number;
  /** @nullable */
  lastMessage?: string | null;
}

export interface ConversationInput {
  workspaceId: number;
  visitorName: string;
  visitorEmail: string;
  sourcePage: string;
  browser?: string;
}

export interface ConversationStatusUpdate {
  status: string;
}

export interface AssignInput {
  agentId: number;
}

export interface Message {
  id: number;
  conversationId: number;
  sender: string;
  senderType: string;
  content: string;
  seen: boolean;
  createdAt: string;
}

export interface MessageInput {
  content: string;
  senderType: string;
  sender?: string;
}

export interface Faq {
  id: number;
  workspaceId: number;
  question: string;
  answer: string;
  keywords: string[];
  createdAt: string;
}

export interface FaqInput {
  question: string;
  answer: string;
  keywords: string[];
}

export interface FaqUpdate {
  question?: string;
  answer?: string;
  keywords?: string[];
}

export interface CannedResponse {
  id: number;
  workspaceId: number;
  title: string;
  content: string;
  createdAt: string;
}

export interface CannedInput {
  title: string;
  content: string;
}

export interface CannedUpdate {
  title?: string;
  content?: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface BotVsAgentCount {
  type: string;
  count: number;
}

export interface AnalyticsSummary {
  totalToday: number;
  totalThisWeek: number;
  /** @nullable */
  avgFirstResponseMs: number | null;
  resolutionRate: number;
  byStatus: StatusCount[];
  botVsAgent: BotVsAgentCount[];
}

export interface AgentActivity {
  agentId: number;
  name: string;
  email: string;
  conversationCount: number;
  isOnline: boolean;
}

export interface Plan {
  id: string;
  name: string;
  price: number;
  features: string[];
}

export interface CheckoutInput {
  successUrl: string;
  cancelUrl: string;
}

export interface CheckoutResponse {
  url: string;
}

export interface PortalResponse {
  url: string;
}

export type ListConversationsParams = {
  status?: string;
};