# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SupportDesk API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register admin + workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Agent/admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /auth/invite:
    post:
      operationId: inviteAgent
      tags: [auth]
      summary: Invite agent by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Agent invited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ── Conversations ─────────────────────────────────────────────────────────
...
  /faqs:
    get:
      operationId: listFaqs
      tags: [faqs]
      summary: List FAQs for workspace
      responses:
        "200":
          description: FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
    post:
      operationId: createFaq
      tags: [faqs]
      summary: Create a FAQ entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /faqs/{id}:
    patch:
      operationId: updateFaq
      tags: [faqs]
      summary: Update a FAQ entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
    delete:
      operationId: deleteFaq
      tags: [faqs]
      summary: Delete a FAQ entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Canned Responses ──────────────────────────────────────────────────────
  /canned:
    get:
      operationId: listCanned
      tags: [canned]
      summary: List canned responses
      responses:
        "200":
          description: Canned responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CannedResponse"
    post:
      operationId: createCanned
      tags: [canned]
      summary: Create canned response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CannedInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CannedResponse"
  /canned/{id}:
    patch:
      operationId: updateCanned
      tags: [canned]
      summary: Update canned response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CannedUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CannedResponse"
    delete:
      operationId: deleteCanned
      tags: [canned]
      summary: Delete canned response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Analytics ─────────────────────────────────────────────────────────────
...
  /billing/checkout:
    post:
      operationId: createCheckout
      tags: [billing]
      summary: Create Stripe checkout session for Pro plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, workspaceName]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        workspaceName:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    InviteInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, name, email, role, workspaceId, isOnline, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        workspaceId:
          type: integer
        isOnline:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # ── Workspace ─────────────────────────────────────────────────────────
...
    Faq:
      type: object
      required: [id, workspaceId, question, answer, keywords, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        question:
          type: string
        answer:
          type: string
        keywords:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    FaqInput:
      type: object
      required: [question, answer, keywords]
      properties:
        question:
          type: string
        answer:
          type: string
        keywords:
          type: array
          items:
            type: string
    FaqUpdate:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
        keywords:
          type: array
          items:
            type: string

    # ── Canned Responses ──────────────────────────────────────────────────
    CannedResponse:
      type: object
      required: [id, workspaceId, title, content, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    CannedInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
    CannedUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string

    # ── Analytics ─────────────────────────────────────────────────────────
...
    CheckoutInput:
      type: object
      required: [successUrl, cancelUrl]
      properties:
        successUrl:
          type: string
        cancelUrl:
          type: string
    CheckoutResponse:
      type: object
      required: [url]
      properties:
        url:
          type: string
...