# lib/api-zod/src/generated/types/cannedResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupportDesk API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CannedResponse {
  id: number;
  workspaceId: number;
  title: string;
  content: string;
  createdAt: Date;
}